requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Werkzeug==3.0.1
python-dotenv==1.0.0
Pillow==10.1.0
firebase-admin==6.4.0
boto3==1.34.0
botocore==1.34.0
requests==2.31.0
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-socketio==5.10.0
Flask-Limiter==3.5.0
redis==5.0.1